[#520] Fix bug introduced that previous commit where setting $ztrap doesn't set $ztrap
The commit merged by !720 introduced a new bug where setting $ztrap didn't set $ztrap if the memory address where the mval $ztrap is supposed to be set to does not overlap in memory with the frame_pointer. The fix undoes the original #520 commit and replaces it with a new one that stores a copy of the string that $ztrap is to be set to in a buffer on the stack and then copies the buffer into the string pool before setting $ztrap correctly.