[#589] Flush buffered IO writes inside external call when YottaDB process output is piped
Background
-----------
This issue was noticed as part of writing a test case for
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/664.

YottaDB has different behavior when stdout is piped and when it isn't! When stdout is not piped,
the `Hello, world!` string shows up. But when stdout is piped, the string does not show up as can be
seen below.

```sh
$ cc -c -fPIC -I$ydb_dist helloworld.c
$ cc -o helloworld.so -shared helloworld.o
$ yottadb helloworld.m
$ LD_LIBRARY_PATH=".:$LD_LIBRARY_PATH" ydb_xc_c=TC01.xc ydb_routines=". $ydb_routines" yottadb -run helloworld
Hello, world!
$ LD_LIBRARY_PATH=".:$LD_LIBRARY_PATH" ydb_xc_c=TC01.xc ydb_routines=". $ydb_routines" yottadb -run helloworld | tee output.txt
$
```

Below are the contents of the relevant external call file and the corresponding C and M programs
needed to reproduce the above behavior.

```m
$ cat helloworld.m
 DO &c.helloworld
 QUIT

$ cat helloworld.xc
libhelloworld.so
helloworld: void hello_world()
```

```c
$ cat helloworld.c

void hello_world() {
    puts("Hello, world!");
}
```

Fix
----
The reason for no output in the piped case is that stdout is opened by YottaDB as file descriptor
1 (stdout), using unbuffered IO, at process startup but the external call writes to the same file
descriptor using `puts()` call which does buffered IO to it. There is a mixing of unbuffered and
buffered IO writes to the same file descriptor by the M and C sections of the same process. And when
YottaDB is about to terminate, it closes all open IO devices without realizing the `puts()` done by
the external call is not yet flushed. Later as part of process exit handling, the operating system
tries to do the flush but it notices the file descriptor 1 has already been closed and the flush
fails with an `EBADF` error code.

While in general it is not a good idea to mix unbuffered and buffered IO, in this particular case,
YottaDB could have been more user friendly by doing a `fflush(NULL);` before it started closing any
of its open file descriptors. That would ensure any writes done by non-YottaDB code (e.g. external
calls etc.) in the same process do get flushed before any file descriptors tied to that buffered IO
structure get closed by YottaDB.

In YottaDB code, the convention is to use `FFLUSH` whenever `fflush` usage is desired. Hence the
change is to add a `FFLUSH(NULL);` call to `io_rundown()` before closing any open file descriptors.