[#592] ps should identify JOB'd process with the same name as the parent
The main change is to `sr_unix/ojstartchild.c` to check if the base program `invocation_mode`
is `MUMPS_RUN` or `MUMPS_DIRECT`. If so, we use the original executable name (based on a new
global variable `invocation_exe_str` if it is non-NULL) as the executable for the jobbed child.

There was one issue with this as `argv[0]` gets changed by `dlopen_libyottadb()` before invoking
`gtm_main()`. The invocation happens with symlinks resolved which defeats the requirement for #592
to track the command line argument specified by the user as is (e.g. "mumps" or "yottadb").

For this purpose, `sr_unix/dlopen_libyottadb.c` was changed to not tamper with `argv[0]` in case
the invocation is for `gtm_main()`. But this tampering is needed for a later call to `ydb_chk_dist()`
and so that call was changed to use the symlink-expanded full path name (obtained from `realpath()`)
after having noted down the untampered `argv[0]` in a new global variable `invocation_exe_str`.