Add RTLD_GLOBAL flag to dlopen() of a YottaDB-plugin (avoids "undefined symbol" errors)
Description
------------
This is an issue that was encountered by Stefan Traby and the fix was provided by him.

Below is his description of the issue.

1) I'm working on a perl-plugin for YottaDB.
2) I got the proof-of-concept up and running quite fast by using the YDBPosix-plugin as build-template.
3) What I didn't get right without a YottaDB-patch is the perl "Dynaloader" which loads .so-libs at runtime.
   It issued the following error.

   ```sh
   /usr/local/lib/yottadb/r129/yottadb: symbol lookup error: /usr/lib/x86_64-linux-gnu/perl/5.28/auto/Fcntl/Fcntl.so: undefined symbol: Perl_xs_handshake
   ```

Fix
----
The fix is to add `RTLD_GLOBAL` in the `dlopen()` in `fgn_getpak()` function (defined in `sr_unix/fgn_getinfo.c`).

The `dlopen()` has been that way since day one and I don't see any reason why `RTLD_GLOBAL` would make it
worse. Given that it helps in at least one known case now, it is considered okay to add it.

Pasted below are the `dlopen()` in the entire code base BEFORE this commit. All of them already have
`RTLD_GLOBAL` except for 2. One of them is being fixed in this commit. The other is in
`sr_linux/hugetlbfs_overrides.c`. In that case, we are interested in only 2 symbols and we do `dlsym()`
calls for them right away so there is no need of `RTLD_GLOBAL` seen at this point.

```c
./sr_linux/hugetlbfs_overrides.c:	handle = dlopen(lpath, RTLD_NOW);
./sr_unix/fgn_getinfo.c:	ret_handle = dlopen(lpath, RTLD_LAZY);
./sr_unix/gtm_zlib.c:	handle = dlopen( lpath, ZLIB_LIBFLAGS | RTLD_MEMBER);
./sr_unix/gtmcrypt_util.h:	handle = dlopen(NULL, (RTLD_NOW | RTLD_GLOBAL));								\
./sr_unix/gtm_tls_loadlibrary.c:	if (NULL == (handle = dlopen(libpath, RTLD_GLOBAL | RTLD_NOW)))
./sr_unix/gtmcrypt_entry.c:	handle = dlopen(&resolved_libpath[0], RTLD_NOW | RTLD_GLOBAL);
./sr_unix/dlopen_libyottadb.c:	handle = dlopen(&libyottadb_realpath[0], (RTLD_NOW | RTLD_GLOBAL));
```