[#594] Fix ydb_incr_s() to buffer lvn value in stringpool as caller buffer is not guaranteed to persist
lvn tree should not have any nodes with values pointing to SimpleAPI caller buffers as they are not
guaranteed to persist. If those buffers are later freed, the lvn tree will be pointing to garbage
memory resulting in SIG-11 or incorrect results.

We noticed the SIG-11 in the `go/randomWalk` subtest (very rare as it happened once in hundreds of
test runs).  The incorrect result part is easily reproducible with a `ydb_dbglvl` env var of `0x40`
and is exercised by the `r130/ydb594` subtest.