[#587] A way to set Initial Value of TEST
New shell environment variable ydb_dollartest is added to YDB

Added new variable dollar_test_default and updated value of dollar_test
based on the prior

Gbldefs variable is used so that dollar_test_default is accessible
process wide.

dm_start.s files are changed so that the dollar_test_default value is
used to set the value of dollar_truth.

Because of the dm_start.s changes the following scenarios will all have
$TEST value initialized based on the new shell environment variable
ydb_dollartest . Regular invocation seen in start of YDB, trigger
invocation seen when there is a trigger set in ydb process and SIMPLE
API invocation seen when ydb is accessed using simpleapi interface using
ydb_ci, ydb_cip.

When ydb_dollartest set with (0 or 1) the value of $TEST will be (0 or 1
respectively).

When the value of ydb_dollartest is unset the value of $TEST will remain
as 1.