[#607] Set flush_trigger_top correctly when auto-upgrading databases from R1.22 or R1.24
This commit ensures that flush_trigger_top is set when a R1.22 or R1.24 database is auto-upgraded. When a database is upgraded from R1.22 or R1.24 to R1.26 or R1.28 format, flush_trigger_time would not be set because GDSMV63007 and GDSMR122 are considered to be the same csd->db_minorver which means that YottaDB skipped the auto-upgrade for upstream version V6.3-007. The change fixes this by doing the initialization for flush_trigger_top if it is set to 0, meaning that it was not set by a previous upgrade.