[YDB#609] Fix issue where MUPIP STOP can cause DBMBPINCFL/DBMBMINCFRE integrity error when instance is frozen
Issue Background
-----------------
We saw a test failure in in-house testing in the `v60000/gtm4525b` subtest (which induces instance
freeze) twice in the past 2 years of testing hundreds of times on dozens of machines. The failure
symptom was that the database had just one DBMBPINCFL integrity error at the end of the test.

```
Integ of region DEFAULT

Block:Offset Level
%YDB-W-DBMBPINCFL,
    2200:0     FF  Master bit map incorrectly marks this local map full
```

There was not much to go with from the test failure outputs so after a thorough review of the code,
an issue was found that can explain the failure symptoms.

This is an edge case in the instance freeze logic that can cause a DBMBPINCFL integrity error if the
instance is frozen and a MUPIP STOP is sent to a process at just the wrong time (after Line 192 but
before Line 206 of `sr_unix/t_end_sysops.c` below).

**sr_unix/t_end_sysops.c**
```c
171 void fileheader_sync(gd_region *reg)
172 {
  .
191         high_blk = cnl->highest_lbm_blk_changed;
192         cnl->highest_lbm_blk_changed = GDS_CREATE_BLK_MAX;      /* Reset to initial value */
  .
206         DB_LSEEKWRITE(csa, udi, udi->fn, udi->fd, 0, (sm_uc_ptr_t)csd, flush_len, save_errno);
  .
212         return;
```

Turns out that inside the `DB_LSEEKWRITE` macro it is possible for the process to terminate if the
instance is frozen and the process had already been sent a SIGTERM (or SIGINT or any other what we
call fatal signal) and the signal handling had been deferred because we were holding the database
critical section trying to finish off something important (e.g. flushing the db file header and/or
master map from shared memory to disk).

The issue is in line 192 which resets the shared memory value of `cnl->highest_lbm_blk_changed`
right after copying it to a local variable in line 191. But if the `DB_LSEEKWRITE` macro in line 206
terminates the process without writing to the db file header, we have effectively not flushed the master
map but have lost the crucial information of how much to flush the master map. Because we reset it to
`GDS_CREATE_BLK_MAX`, the next process that invokes `fileheader_sync()` will see the initial value
in the shared memory copy of `cnl->highest_lbm_blk_changed` and incorrectly decide there is nothing
to flush in the master map when actually there is. This can explain the DBMBPINCFL integrity error.

As to how the `DB_LSEEKWRITE` can exit the process in the middle of the macro, below is a macro call
graph pasting the relevant code path.

**sr_port/anticipatory_freeze.h**
```c
339 #define DB_LSEEKWRITE(CSA, UDI, DB_FN, FD, OFFSET, BUFF, SIZE, STATUS)
  .
349         DO_LSEEKWRITE(CSA_LOCAL, DB_FN, FD, OFFSET, BUFF, SIZE, STATUS, fake_db_enospc, DB_LSEEKWRITE_HANG(CSA),
350                         LSEEKWRITE_IS_TO_DB);
```

```c
409 #define DO_LSEEKWRITE(CSA, FNPTR, FD, OFFSET, BUFF, SIZE, STATUS, FAKE_WHICH_ENOSPC, HANG, LSEEKWRITE_TARGET)
  .
416         if (NULL != local_csa)
417                 WAIT_FOR_REPL_INST_UNFREEZE_SAFE(local_csa);
```

```c
212 #define WAIT_FOR_REPL_INST_UNFREEZE_SAFE(CSA)           \
  .
219         if (IS_REPL_INST_FROZEN_JPL(local_jnlpool))     \
220                 WAIT_FOR_REPL_INST_UNFREEZE(CSA);       \
```

```c
  204 #define WAIT_FOR_REPL_INST_UNFREEZE(CSA)
  205                 wait_for_repl_inst_unfreeze(CSA)
```

```c
508 static inline void wait_for_repl_inst_unfreeze(sgmnt_addrs *csa)
  .
526                 WAIT_FOR_REPL_INST_UNFREEZE_NOCSA_JPL(local_jnlpool);
```

```c
230 #define WAIT_FOR_REPL_INST_UNFREEZE_NOCSA_JPL(JPL)                                      \
231                 wait_for_repl_inst_unfreeze_nocsa_jpl(JPL)
```

```c
477 static inline void wait_for_repl_inst_unfreeze_nocsa_jpl(jnlpool_addrs_ptr_t jpl)
478 {
  .
484         while (jpl->jnlpool_ctl->freeze)
485         {
486                 if (exit_state != 0)
487                 {
488                         forced_exit_err_display();
489                         EXIT(-exi_condition);
490                 }
```

Fix
---
The fix is to move line 192 to just before line 212 above. That is, do not reset the shared memory
copy until the `DB_LSEEKWRITE` succeeds. This way even if the `DB_LSEEKWRITE` macro terminates this
process, some other process can see the shared memory value and update the master map on disk correctly.