[#597] Save plugin build directories if building the plugin fails
This commit modifies the ydbinstall script to save build directories for plugins if building the plugin failed. Previously, the install script would delete the plugin build directory whether or not the plugin was successfully built. The change is to get the exit code of the final command related to building the plugin. If it is 0, then the plugin build directory will be deleted. Otherwise, the plugin build directory will be saved. This change has been tested by simulating build errors for each of the plugins and verifying that the build directories from the failed build were retained.