[#581] $ZPARSE() option to follow symbolic links
This change is required to ensure that ZPARSE can handle symbolic links,
and fetches names of files that are symbolically linked and not the link
itself.

The code makes changes in op_fnzparse.c and parse_file.c files where in
the first file, the change adds a new argument for the fifth
parameter for zparse called SYMLINK. When the argument is SYMLINK yottadb
searches for symbolic link.

The parse_file has logic for searching for symbolically linked file. By
using canonicalize_file_name and recursively calling parse_file it
searches for symbolically linked files.