Overwrite pkg-config file in ydbinstall.sh if YDB version is the same
Previously, the pkg-config file would not be overwritten if the versions
were the same, only if the previous version was strictly less.

That meant if V999_R129 was installed and we now try to install V999_R131, the pkgconfig file would not be overwritten since both of them have the same ydb_version of r9.99.

This change does not affect production environments,
which will never have two versions of YDB with the same version number
but a different pkg-config file.