[V63009] Fix imptp related test failures caused by a conflict missed by git
In the V6.3-009 changes, a block of code in sr_unix/gds_rundown.c that was previously removed from YottaDB's code base as part of the ydb150 changes was moved. This was not picked up as a conflict by git and the code was treated as a new addition that did not conflict with existing YottaDB changes to that file. This commit resolves this conflict and fixes the test failures by removing this block of code from sr_unix/gds_rundown.c.