[V63009] Fix warning in sr_unix/gtmcrypt_util.c by changing a strncpy to a memcpy
This commit fixes the compiler warning `sr_unix/gtmcrypt_util.c:420:5: warning: 'strncpy' output truncated before terminating nul copying as many bytes from a string as its length [-Wstringop-truncation]` by changing the strncpy to a memcpy. This is equivalent because the value of len is guaranteed to be strlen(lpasswd) and env_value is guaranteed to be malloced for len + 1 bytes.