[V63009][#652] Fix V6.3-009 bug where some base conversions from hexadecimal to octal were incorrect
The v63005/gtm5574 test fails around 2% of the time in internal testing with the V6.3-009 changes integrated. This failure has also been seen on versions V6.3-009 and V6.3-013 of the upstream software. This failure always involves 19 digit decimal numbers that are exactly 15 digits in hexadecimal form and which are too large to be stored precisely as a numeric mval in YottaDB (specifically numbers between 1000000000000000001 and 1152921504606846975 in decimal or DE0B6B3A7640001 and FFFFFFFFFFFFFFF in hexadecimal). One of the V6.3-009 changes to %HO does the base conversion by converting the number back to decimal and then to octal using the modular and integer division functions as this is faster than converting directly from base 16 to base 8. This commit fixes the bug by converting 15 digit hexadecimal numbers directly to octal.