[V63009] Fix commit-verify pipeline failure by skipping the copyright check for "merge with conflicts" commit
This commit modifies the commit-verify pipeline so that the copyright check will not run on a `merge GT.M Vx.x-xxx into YottaDB mainline (with conflicts)` commit which will prevent the copyright check from failing because YottaDB copyrights were not added to files that have only been changed by the upstream software. This was necessary because the pipeline was failing on this commit for the V6.3-009 merge due to several such spurious "missing copyrights" that the pipeline was finding.