[V63009] Address review comments
This commit addresses review comments on the V6.3-009 merge code changes. There are 3 changes. A typo in Steve Estes's name has been fixed in ci/commit_verify.sh, a comment in sr_port/op_zstep.c has been changed because of a suspected typo and the OBJ_UNIX_LABEL has been increased from 7 to 8. The OBJ_UNIX_LABEL was increased because of the addition of the $ZMLKHASH ISV. If OBJ_UNIX_LABEL were not increased, this could cause issues with programs that call ISVs that come after $ZMLKHASH in alphabetical order and which were compiled on an old version of YottaDB.