[#654] Fix ydbinstall failures on AArch64 machines running Ubuntu 20.04 due to "missing" libtinfo.so.5
On Ubuntu 20.04, ydbinstall currently requires libtinfo.so.5 to successfully install YottaDB. However, the script always looks for it at the path where it will be installed on x86_64 (/lib/x86_64-linux-gnu/libtinfo.so.5) even on AArch64 machines where it is installed at a different location (/lib/aarch64-linux-gnu/libtinfo.so.5). This commit fixes this by looking for the correct location for the architecture on which ydbinstall is being run.

As only Raspbian and Debian are currently supported by YottaDB on ARMv6l and ARMv7l, this commit does not address the case where a user attempts to install YottaDB on Ubuntu 20.04 on those architectures. This change likely eliminates the check for libtinfo.so.5 on those platforms altogether but this has not been tested.