Fix SIG-11 in sr_port/iosocket_connect.c due to errlen not being initialized
A configuration issue on a machine that caused test failures in internal testing exposed a SIG-11 in sr_port/iosocket_connect. The test produced a core file where errptr was set to "No route to host" but errlen was set to an extremely high value. Upon investigation, we discovered that the errlen here was only set within the SET_DOLLARDEVICE_ONECOMMA_ERRSTR() macro which only ran if iod->state was equal to dev_open. If it wasn't and there was an error with IO, then errlen would be uninitialized when rts_error_csa was called and this could result in a SIG-11 while YottaDB attempted to print the error message. This commit fixes this by creating a macro called SET_ERRPTR_AND_ERRLEN() to ensure that errptr and errlen are initialized correctly and consistently.