Fix Dockerfile caching
Previously, each time any file in the whole source repository was
modified, all of YDB had to be rebuilt, along with any intermediate
layers. Now, changing files (such as the Dockerfile itself) that do not
affect the build does not require recompiling YDB.

While working on this, I found that the entire build directory was being
sent to the docker daemon, even though it wasn't needed for the build. I
added it `.dockerignore`, reducing the build time substantially.