[#591] Remove ifdef TCP_NODELAY from YottaDB code
It was thought that TCP_NODELAY was not defined in the YottaDB code
base, and thus the sections that used it were disabled. Upon further
investigation, it was found that TCP_NODELAY is set by the Operating
System, and is always set to be enabled. It is now determined that
TCP_NODELAY is a historical flag, and has no meaning today for the
platforms that YottaDB supports. The change made just removes the flag
from the YottaDB code base, as the code the flag protects is always
enabled.