Add additional information to SYSCALL error from ss_context_mgr
A weekend test (v53003/D9I10002703) failed with a SYSCALL error during shutdown indicating that it couldn't delete the
shared memory associated with a snapshot. The error code is insufficient to understand what the error is/was:

> %YDB-E-SYSCALL, Error received from system call Error with shmdt -- called from module /Distrib/YottaDB/V999_R131/sr_unix/ss_context_mgr.c at line 174,%SYSTEM-E-ENO22, Invalid argument
>            At M source location child+1^d002703

What we need to know to understand this error is the actual shared segment address being deleted. This change adds that
information with an additional ERR_TEXT error that contains "Shared buffer addresss: 0x[hhhhhhhh]hhhhhhhh"