First set of changes for [#674] - Cross references for complete nodes
Creates triggers to maintain cross references and computes cross references
for a global variable at a specified subscript level.

Usage: $$XREFDATA^%YDBAIM(gbl,nsubs,sep,pnum,nmonly)
Parameters:
- gbl is the global variable name, e.g., "^ABC"
- nsubs is the number of subscripts at the level for which the xref is desired
- sep is the piece separator for values at that node; if unspecified or the
  empty string, xref is for entire node value
- pnum is a comma-separated list of piece numbers for which cross references
  should exist; ignored for xrefs of entire node values
- nmonly, if 1, means just return the xref global variable name but don't set
  triggers or compute xrefs
Return value: name of global variable with cross reference
e.g., "^%ydbAIMDZzUmfwxt80MHPiWLZNtq4"

$data($$XREFDATA^%YDBAIM(gbl,nsubs,sep,pnum,1))=11 means triggers are in place
and cross referencing is complete

Since the function uses $PIECE(), sep is interpreted according to whether the
process is in M mode or UTF-8 mode.  While cross references for entire nodes
are the same for M and UTF-8 modes, there are separate cross references for
pieces of nodes.  Cross references are in global variables such as
^%ydbAIMDZzUmfwxt80MHPiWLZNtq4, which are derived from a prefix of "^%ydbAIMD"
and a suffix of a 128-bit hash of the variable name, number of subscripts and
piece separator, as well as $zchset if the piece separator is not the empty
string, rendered using an alphabet of 64-characters to create a legal YottaDB
global variable name.  The subscripts of cross reference variables are:
- (pnum,value,sub[,sub]) where pnum is a positive integer for cross references
  of pieces of nodes; and
- (value,sub[,sub]) for cross references of entire nodes.

A limitation of this technique is that the values being cross referenced plus
the subscripts must have a total length within the limit of YottaDB global
keys (currently 1019 bytes). Another is that $char(10) is problematic as
the piece separator.

The approach is to first create triggers to maintain cross references, and
then scan the global variable to generate cross references. This way, any
global variable changes will have cross references maintained by the triggers,
and when returning to caller, the cross reference is complete, and will remain
Consistent thereafter.  ^%ydbAIMD* global variables should be mapped to a
region with NULL_SUBSCRIPTS set to EXISTING or ALWAYS. Although no xrefs use
null subscripts, calls to $data() will use null subscripts if $ZTOLDVAL is the
empty string and need to succeed.

Just as cross references are metadata for application global variables,
metadata for the cross reference global variables are stored in the xref
global variables as follows:
- root node is the application global variable
- (0) node is $zut, a space, and $zyrelease
- (1) node is the number of subscripts
- (2) node is the piece separator, "" for an xref of entire nodes
- (3) node is the piece numbers, in the format used for specifying triggers
- (4) node and up are triggers for this cross reference
Since the cross references themselves must have at least two subscripts, any
node with one subscript is metadata for the cross reference.