[#680] Fix issue with fileheader-sync timer not found
The relink_0/barrage test failed with an assert failure in gt_timers.c cause by its inability to
find a timer that it was trying to cancel because a real timer interrupt that was not blocked got
in and removed the timer.

The issue https://gitlab.com/YottaDB/DB/YDB/-/issues/680#description has the full write up on it.