[#685] Fix handling of invalid input for %OH and %OD
This commit fixes the handling of invalid input by %OH and %OD so that it is handled consistently with how %DH, %DO, %HD and %HO handle invalid input: Convert the input until the first invalid character and then ignore all remaining characters. This is tested by the v63009/gtm9115 subtest.