Added missing Ubuntu dependencies to README.md
The packages `make`, `gcc` and `curl` are required for building and installing YDB on Ubuntu 20.04, but they are not listed in the list of required dependencies for this platform in `README.md`. As a result, following those instructions as they are lead to build and installation errors. This is now corrected by listing these missing dependencies in that file.

For the record, the absence of these packages led to CMake errors due to there being no C compiler and no way to process "Unix Makefiles", i.e. no `gcc` or `make`. The absence of `curl` caused the retrieval of YDBPosix in `ydbinstall` to fail, as this script relies on `curl` for downloading the source tarball.

Note that `file` and `git` were also added to this list to match the list of packages installed in the Dockerfile. Also, the requirements lists for Arch and CentOS/RHEL were updated to include these dependencies as well.