[#673] Change M Lock hash from 32 bits to 64 bits, fix possible M Lock failure, increase default lock size from 40 to 220 pages
The current M lock hash is 32 bits but this has been known to cause "full neighborhood" conditions
in the hopscotch hash used for M Locks and causing locking failures with a zero timeout. The existing
[V63009] M Lock code base will rehash (bumping the key seed) or resize the hash table (doubling it)
to get a better distribution but it only does this once. With a 64 bit hash, the distribution is better
in the first place providing a method of avoiding the issue.

While making changes to tests for the 32->64 bit hash size increase, the locks/lkeexact test started
throwing a core in gtm_malloc() due to storage corruption. After a long chase, the issue was found in
how mlk_pvtblk_create() allocates the pvtblk and the buffer for subscript mval strings that follows it.

The length of the pvtblk was being incorrectly computed in the MLK_PVTBLK_SIZE macro as far back as
the V5 versions we have. We are not sure why these changes brought the issue to light but correcting the
size cured the array overrun that was occurring that obscured the trailing validation marker. But the
test failed every time until mlk_pvtblk_create.c was fixed.

The increase of the hash from 32 to 64 bits caused about a 10% increase in lock usage which the user
should be aware of. To this end, we have also increased the default lock size from 40 to 220 "pages"
of 512 bytes each. But users with existing lock tables that are running close to the edge will want to
give their lock space value a 10-15% overall bump to avoid problems. In addition, when we first
increased the hash from 4 to 8 bytes (32->64 bits), the number of shrblks went down disproportionately
to the other resources. This is because the shrblk itself grew while the others did not. To counter this,
a minor minor change was made to the floating point value shrblk_adj so it weighs mlk_shrblks to be just
slightly more important to allocate than they were so the changes in the lock resources reported by
LOCKSPACEINFO were more in proportion.

Now that the YottaDB fixes for #438 were reverted as part of prior commits (to keep the YottaDB M-lock
code base as close as possible to the GT.M M-lock code base) it was necessary to retrofit #438 back in
to avoid the situation where an M-lock command with a 0 timeout returns with $test set to 0 even though
the lock is unowned. Towards that, op_lock2.c has been changed to check if mlk_lock() returned with a
gc_needed, rehash_needed, or resize_needed flag set and if so it retries the mlk_lock() (even if the
lock timeout is 0) for a maximum of two times per lock per process.

Note the DefinedTypes files were rebuilt due to the changes in lock structures.