[#697] Fix bug in CONVNEG^%CONVBASEUTIL for octal numbers introduced by V6.3-009 changes
This commit fixes a bug in CONVNEG^%CONVBASEUTIL that returned incorrect two's complement conversions for octal numbers where the first 2 digits of the number are non-zero and any remaining digits are zero. This function is used by %DO to convert negative decimal numbers to a two's complement octal number. The bug was discovered when the v63005/gtm5574 test failed in internal testing after the 5 digit negative number section of the test randomly generated a number that triggered this bug causing a very rare test failure that could not be reproduced in 1000 attempts. This bug is fixed via a hybrid of the r1.30 version and V6.3-009 versions of this function.