[V63010] Fix relink test failures caused by V6.3-010 changes to zlink
The V6.3-010 changes conflicted with previous YottaDB changes from issue #432, effectively undoing these changes in part by changing the structure of `op_zlink.c`. This caused the relink subtests basic, barrage, recursive and truncate to fail 100% of the time. These changes restored NOTMNAME error messages for zlink commands ending with `.m` or `.o`. This commit ensures that the `.m` or `.o` is still automatically removed by YottaDB. This also ensures that commands like `ydb -o=y.o.m x.m` work correctly (by ensuring that the check for `.o` and `.m` endings does not execute twice within `zl_cmd_qlf.c`) and fixes a bug in the original #432 changes. This bug involved filenames ending with an `m` or an `o` where no extension was specified and where the second to last character was not a `.`. Such filenames were incorrectly truncated which led to a failure in the `v60001/gtm7461` test due to `datinfo` being truncated to `datinf`.