[V63010] Some changes to bring all 3 flavors of obj_filesp.c into line (sr_x86_64, sr_armv7l, sr_aarch64)
The differences between the upstream version of obj_filesp.c (for x86-64) are significantly
different from the YottaDB version. YottaDB had already done the object file changes needed
for the object files to work in the linker/loader (and various ELF utilities like eu-elflint
from the elfutils package on Debian) but that happened in V63010 for the upstream version
so caused some significant module churn. Most of that churn was picked up in the original
collision fixes resolved during the merge of the V63010 base but the few that escaped are
picked up here.

Also, while researching this, it was  discovered that the ehdr->e_ident[EI_OSABI] field (which
YottaDB had been filling in with ELFOSABI_LINUX and which YottaDB had been filling in with
ELFOSABI_SYSV) was supposed to be filled in with ELFOSABI_NONE unless the object file was
cross-compiled from another system so that was changed in all three versions as well.

The original YottaDB issue (#496) changes where this additional object support was done
mentioned that after that work, there was still a minor error coming out of eu-elflint but
with changes done subsequently to not hard-code the section IDs but to have the elf routines
assign them and others, we now see NO ERRORs out of eu-elflint on object files created by YDB.
However, we do still see object file errors coming out of eu-elflint for object files coming
out of the upstream V63010 software.