[YDBTest#349] Fix incorrect reference to "XCCONVERT" message in help file
This commit fixes an incorrect reference to an `XCCONVERT` error message in a help file. The only references to `XCCONVERT` are in this help file and in the upstream release notes for V6.3-009. It remains in their help files as of V7.0-000. The error message is actually called `ZCCONVERT`. This fixes this upstream error which we discovered while while working on the v63009/gtm9149 test.