Enhance ydbinstall.sh to display user-friendly error messages when a plugin build fails due to download issues
This commit modifies the ydbinstall script to address how the plugin install options --posix --zlib and --octo handle unsuccessful downloads. All 3 of these options rely upon downloading the current version of the plugin from gitlab. Should it be unable to download the plugin, it currently continues through the build process producing several error messages in a row. This commit changes this by detecting if the download fails and, if it fails, it displays a user-friendly error message informing the user that the download failed.