[#661] ydb_env_set creates a three-region database by default
The default database specified by `gdedefaults.gtc` has three regions with corresponding segments, in addition to the DEFAULT region which was the sole region previously:

- YDBOCTO with all `^%ydbOcto*` global variables (with `"Octo"` being case-insensitive - i.e., all 16 variations of the name) mapped to database file `$ydb_dir/$ydb_rel/g/%ydbocto.dat`. It has before image journaling to a journal file of `$ydb_dir/$ydb_rel/g/%ydbocto.mjl`. Other database properties are identical to DEFAULT.

- YDBAIM with all `^%ydbAIM*` global variables (with `"AIM"` being case insensitive - i.e., all 8 variations of the name) mapped to database file `$ydb_dir/$ydb_rel/g/%ydbaim.dat`. The region has AutoDB set, with no journaling. The segment has MM access method, a 1KiB block size, key size of 992. The large key size is because metadata includes data in the subscripts, and the small block size is because node values are either empty strings or integers, so nodes will be small and there will be more nodes per Kilobyte than in typical data.

Sourcing `ydb_env_set` recovers the YDBOCTO region just as it recovers the DEFAULT region, but deletes the YDBAIM region if that was not cleanly shut down, as the data is metadata that is (re)generated when needed by calling `$$XREFDATA^%YDBAIM()` / `ydb_aim_xrefdata_s()`. Changing the access method to BG, and enabling & turning on before-image journaling prior to an unclean shutdown results in `ydb_env_set` recovering the YDBAIM region just as it does the other regions.

Additionally:

- The default mode is UTF-8 if YottaDB was installed with UTF-8 support. Previously, it was M mode by default even if UTF-8 mode support was installed.

- For UTF-8 mode, it checks whether a locale is set. If not, it uses the first UTF-8 locale in the `locale -a` output, and terminates with an error if one is not found.

- In case of error, location of error file is output. Previously, it was dumped on the shell session.

- It handles the case where replication is turned on.